requirements.txt
tensorflow>=2.8.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
pillow>=8.2.0
tqdm>=4.60.0
pandas>=1.2.0
seaborn>=0.11.0
fastapi>=0.70.0
uvicorn>=0.15.0
streamlit>=1.0.0
requests>=2.25.0
kaggle>=1.5.12
locust>=2.9.0
python-multipart>=0.0.6
aiofiles>=23.1.0
